Attempted test refactoring with more efficient logging
