+++ setup/drop.js
@@ -2,19 +2,22 @@ import Connection from "../app/db/connection.js";
 import { set } from "../app/util/util.js";
 import schemaData from "./setup.js";
 
-process.stdout.write(
-  set("\n-------------------------------\n[-] DROPPING ALL COLLECTIONS...\n-------------------------------\n").blue
-);
+const silent = process.argv.includes("--silent") || process.argv.includes("-s");
+
+silent ||
+  process.stdout.write(
+    set("\n-------------------------------\n[-] DROPPING ALL COLLECTIONS...\n-------------------------------\n").blue
+  );
 
 const schemas = schemaData.map((data) => data.schema);
 
 try {
-  await Connection.open();
+  await Connection.open(silent);
   await dropIfExists(schemas);
-  await Connection.close();
+  await Connection.close(silent);
   process.exit();
 } catch (error) {
-  process.stdout.write(set(`[-] Error while dropping database: ${error} \n`).red);
+  silent && process.stdout.write(set(`[-] Error while dropping database: ${error} \n`).red);
   process.exit(1);
 }
 
@@ -27,12 +30,13 @@ async function dropIfExists(schemas) {
   for (const schema of schemas) {
     try {
       await schema.collection.drop();
-      process.stdout.write(`${set(`[-] DROP ${schema.collection.name}: `).blue}${set("SUCCESSFUL\n").green}`);
+      silent || process.stdout.write(`${set(`[-] DROP ${schema.collection.name}: `).blue}${set("SUCCESSFUL\n").green}`);
     } catch (error) {
       if (error.code === 26) {
-        process.stdout.write(`${set(`[-] DROP ${schema.collection.name}: `).blue}${set("DOES NOT EXIST\n").yellow}`);
+        silent ||
+          process.stdout.write(`${set(`[-] DROP ${schema.collection.name}: `).blue}${set("DOES NOT EXIST\n").yellow}`);
       } else {
-        process.stdout.write(`${set(`[-] DROP ${schema.collection.name}: `).blue}${set("FAILED\n").red}`);
+        silent || process.stdout.write(`${set(`[-] DROP ${schema.collection.name}: `).blue}${set("FAILED\n").red}`);
         throw new Error(error);
       }
     }
