+++ tests/__tests__/api/user/postLogin.js
@@ -1,0 +1,71 @@
+import { execSync } from "child_process";
+import request from "supertest";
+import app from "../../../../app/app.js";
+import Connection from "../../../../app/db/connection.js";
+import logTestSuite from "../../../util.js";
+import {
+  validUserLoginRaisa,
+  validUserLoginEthan,
+  invalidUserLoginEthan,
+  invalidUserLoginRaisa,
+  pendingUserLoginAce,
+} from "../../../testData/userTestData.js";
+
+afterAll(async () => {
+  await Connection.close(!logTestSuite);
+});
+
+beforeEach(async () => {
+  execSync("npm run reset-s", { stdio: "inherit" });
+});
+
+describe("User Login Tests", () => {
+  test("should login a valid user Raisa", async () => {
+    const response = await request(app).post("/user/login").send(validUserLoginRaisa);
+
+    logTestSuite.user && console.log(response.body);
+    expect(response.status).toBe(200);
+    expect(response.body.message).toBe("Login successful");
+    expect(response.headers["set-cookie"]).toBeDefined();
+  });
+
+  test("should login a valid user Ethan", async () => {
+    const response = await request(app).post("/user/login").send(validUserLoginEthan);
+
+    logTestSuite.user && console.log(response.body);
+    expect(response.status).toBe(200);
+    expect(response.body.message).toBe("Login successful");
+    expect(response.headers["set-cookie"]).toBeDefined();
+  });
+
+  test("should not login with incorrect password for Ethan", async () => {
+    const response = await request(app).post("/user/login").send(invalidUserLoginEthan);
+
+    logTestSuite.user && console.log(response.body);
+    expect(response.status).toBe(400);
+  });
+
+  test("should not login with incorrect password for Raisa", async () => {
+    const response = await request(app).post("/user/login").send(invalidUserLoginRaisa);
+
+    logTestSuite.user && console.log(response.body);
+    expect(response.status).toBe(400);
+  });
+
+  test("should not login with non-existent username", async () => {
+    const response = await request(app).post("/user/login").send({
+      username: "nonexistentuser",
+      password: "password",
+    });
+
+    logTestSuite.user && console.log(response.body);
+    expect(response.status).toBe(404);
+  });
+
+  test("should not login if user is unapproved", async () => {
+    const response = await request(app).post("/user/login").send(pendingUserLoginAce);
+
+    logTestSuite.user && console.log(response.body);
+    expect(response.status).toBe(400);
+  });
+});
